break out ApplicationBookmarkSerializer
