+++ api/serializers.py
@@ -1,4 +1,4 @@
-from core.models.application import ApplicationScore, ApplicationBookmark
+from core.models.application import ApplicationScore
 from core.models.credential import Credential
 from core.models.group import Group
 from core.models.quota import Quota
@@ -30,24 +30,6 @@ from threepio import logger
 
 
 # Serializers
-class ApplicationBookmarkSerializer(serializers.ModelSerializer):
-    """
-    """
-    #TODO:Need to validate provider/identity membership on id change
-    type = serializers.SerializerMethodField('get_bookmark_type')
-    alias = serializers.SerializerMethodField('get_bookmark_alias')
-
-    def get_bookmark_type(self, bookmark_obj):
-        return "Application"
-
-    def get_bookmark_alias(self, bookmark_obj):
-        return bookmark_obj.application.uuid
-
-    class Meta:
-        model = ApplicationBookmark
-        fields = ('type', 'alias')
-
-
 class ApplicationScoreSerializer(serializers.ModelSerializer):
     """
     """
