+++ api/serializers/__init__.py
@@ -13,3 +13,4 @@ from .boot_script_serializer import BootScriptSerializer
 from .application_threshold_serializer import ApplicationThresholdSerializer
 from .application_serializer import ApplicationSerializer
 from .paginated_application_serializer import PaginatedApplicationSerializer
+from .application_bookmark_serializer import ApplicationBookmarkSerializer
